+++ packages/commonwealth/client/scripts/views/components/SublayoutHeader/useJoinCommunity.tsx
@@ -8,9 +8,9 @@ import AddressInfo from 'models/AddressInfo';
 import React, { useState } from 'react';
 import app from 'state';
 import { addressSwapper } from 'utils';
-import { TOSModal } from 'views/components/SublayoutHeader/TOSModal';
 import { AccountSelector } from 'views/components/component_kit/cw_wallets_list';
 import { isWindowMediumSmallInclusive } from 'views/components/component_kit/helpers';
+import TOSModal from 'views/modals/TOSModal';
 import { LoginModal } from 'views/modals/login_modal';
 import { useFlag } from '../../../hooks/useFlag';
 import { AuthModal } from '../../modals/AuthModal';
