+++ packages/commonwealth/client/scripts/views/modals/TOSModal/TOSModal.scss
@@ -1,4 +1,4 @@
-@import '../../shared.scss';
+@import '../../../../styles/shared.scss';
 
 .TOSModal {
   display: flex;
