+++ packages/commonwealth/client/scripts/views/modals/TOSModal/index.ts
@@ -1,0 +1,3 @@
+import TOSModal from './TOSModal';
+
+export default TOSModal;
